+++ middleware.js
@@ -1,3 +1,7 @@
+const ExpressError = require('./utils/ExpressError');
+const { campgroundSchema } = require('./schemas');
+const Campground = require('./models/campground');
+
 module.exports.isLoggedIn = (req, res, next) => {
 	if (!req.isAuthenticated()) {
 		req.session.returnTo = req.originalUrl;
@@ -10,6 +14,26 @@ module.exports.isLoggedIn = (req, res, next) => {
 module.exports.storeReturnTo = (req, res, next) => {
 	if (req.session.returnTo) {
 		res.locals.returnTo = req.session.returnTo;
+	}
+	next();
+};
+
+module.exports.validateCampground = (req, res, next) => {
+	const { error } = campgroundSchema.validate(req.body);
+	if (error) {
+		const msg = error.details.map((details) => details.message).join(',');
+		throw new ExpressError(msg, 400);
+	} else {
+		next();
+	}
+};
+
+module.exports.isAuthor = async (req, res, next) => {
+	const { id } = req.params;
+	const campground = await Campground.findById(id);
+	if (!campground.author.equals(req.user._id)) {
+		req.flash('error', 'Action Not Authorized');
+		return res.redirect(`/campgrounds/${id}`);
 	}
 	next();
 };
