+++ routes/campgrounds.js
@@ -1,20 +1,8 @@
 const express = require('express');
 const router = express.Router();
 const catchAsync = require('../utils/catchAsync');
-const ExpressError = require('../utils/ExpressError');
 const Campground = require('../models/campground');
-const { campgroundSchema } = require('../schemas');
-const { isLoggedIn } = require('../middleware');
-
-const validateCampground = (req, res, next) => {
-	const { error } = campgroundSchema.validate(req.body);
-	if (error) {
-		const msg = error.details.map((details) => details.message).join(',');
-		throw new ExpressError(msg, 400);
-	} else {
-		next();
-	}
-};
+const { isLoggedIn, isAuthor, validateCampground } = require('../middleware');
 
 router.get(
 	'/',
@@ -58,16 +46,12 @@ router.post(
 router.get(
 	'/:id/edit',
 	isLoggedIn,
+	isAuthor,
 	catchAsync(async (req, res) => {
-		const { id } = req.params;
-		const campground = await Campground.findById(id);
+		const campground = await Campground.findById(req.params.id);
 		if (!campground) {
 			req.flash('error', 'Campground Not Found');
 			return res.redirect('/campgrounds');
-		}
-		if (!campground.author.equals(req.user._id)) {
-			req.flash('error', 'You are not authorized to edit');
-			return res.redirect(`/campgrounds/${id}`);
 		}
 		res.render('campgrounds/edit', { campground });
 	})
@@ -76,15 +60,11 @@ router.get(
 router.put(
 	'/:id',
 	isLoggedIn,
+	isAuthor,
 	validateCampground,
 	catchAsync(async (req, res) => {
 		const { id } = req.params;
-		const campground = await Campground.findById(id);
-		if (!campground.author.equals(req.user._id)) {
-			req.flash('error', 'You are not authorized to edit');
-			return res.redirect(`/campgrounds/${id}`);
-		}
-		const camp = await Campground.findByIdAndUpdate(id, {
+		const campground = await Campground.findByIdAndUpdate(id, {
 			...req.body.campground,
 		});
 		req.flash('success', 'Successfully Updated');
@@ -95,6 +75,7 @@ router.put(
 router.delete(
 	'/:id',
 	isLoggedIn,
+	isAuthor,
 	catchAsync(async (req, res) => {
 		const { id } = req.params;
 		await Campground.findByIdAndDelete(id);
